# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: reVoAgent API
  description: Enterprise AI Development Platform with Cost-Optimized Model Management
  version: 1.0.0
  contact:
    name: reVoAgent Support
    url: https://github.com/heinzdev11/reVoAgent
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.revoagent.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/chat:
    post:
      summary: Send chat message
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
  /api/chat/multi-agent:
    post:
      summary: Multi-agent chat session
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiAgentChatRequest'
      responses:
        '200':
          description: Multi-agent responses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiAgentChatResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChatRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
        context:
          type: object
        options:
          type: object
    ChatResponse:
      type: object
      properties:
        response:
          type: string
        model_used:
          type: string
        cost:
          type: number
          format: float
        response_time:
          type: string
        agents_involved:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
    MultiAgentChatRequest:
      type: object
      required:
        - message
        - agents
      properties:
        session_id:
          type: string
        message:
          type: string
        code:
          type: string
        agents:
          type: array
          items:
            type: string
    MultiAgentChatResponse:
      type: object
      properties:
        session_id:
          type: string
        responses:
          type: array
          items:
            type: object
            properties:
              agent:
                type: string
              response:
                type: string
              confidence:
                type: number
                format: float
        consensus:
          type: string
        total_cost:
          type: number
          format: float
        collaboration_score:
          type: number
          format: float
...